% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqm-included-datasets.R
\docType{data}
\name{ant}
\alias{ant}
\title{HLA antibodies}
\format{
A data frame with 10459 rows and 2 variables:
\describe{
\item{ ID }{ an unique identifyer }
\item{ abs }{ HLA antibodies }
}
}
\source{
simK:: Abs_df(candidates = cds)
}
\usage{
ant
}
\description{
data in transplant candidates' HLA-antibodies
}
\keyword{datasets}
