% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqm-included-datasets.R
\docType{data}
\name{cds}
\alias{cds}
\title{transplant candidates}
\format{
A data frame with 2000 rows and 13 variables:
\describe{
\item{ ID }{ an unique identifyer }
\item{ bg }{ blood group }
\item{ A1 }{ HLA-A allele 1 }
\item{ A2 }{ HLA-A allele 2 }
\item{ B1 }{ HLA-A allele 1 }
\item{ B2 }{ HLA-A allele 2 }
\item{ DR1 }{ HLA-A allele 1 }
\item{ DR2 }{ HLA-A allele 2 }
\item{ age }{ years old }
\item{ dialysis }{ months on dialysis }
\item{ cPRA }{ calculated PRA percentage }
\item{ urgent }{ clinicaly urgent }
}
}
\source{
simK::candidates_df(n = 2000)
}
\usage{
cds
}
\description{
data in wait listed transplant candidates
}
\keyword{datasets}
