% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqm_v0-included-datasets.R
\docType{data}
\name{dns}
\alias{dns}
\title{organ donors}
\format{
A data frame with 280 rows and 9 variables:
\describe{
\item{ ID }{ an unique identifyer }
\item{ bg }{ blood group }
\item{ A1 }{ HLA-A allele 1 }
\item{ A2 }{ HLA-A allele 2 }
\item{ B2 }{ HLA-A allele 1 }
\item{ B1 }{ HLA-A allele 2 }
\item{ DR1 }{ HLA-A allele 1 }
\item{ DR2 }{ HLA-A allele 2 }
\item{ age }{ years old }
}
}
\source{
simK::donors_df(n = 280)
}
\usage{
dns
}
\description{
data in donors for organ transplantation
}
\keyword{datasets}
