% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqm_v0.R
\name{eqm_v0}
\alias{eqm_v0}
\title{Equity matrix algorithm}
\usage{
eqm_v0(
  iso = TRUE,
  dABO = "O",
  dA = c("1", "2"),
  dB = c("15", "44"),
  dDR = c("1", "4"),
  donor.age = 60,
  df.abs = cabs,
  data = candidates,
  n = 2,
  q2 = 60,
  q3 = 80,
  uj.matx = uj_matx_v0(),
  check.validity = FALSE
)
}
\arguments{
\item{iso}{A logical value for ABO isogroupal compatibility.}

\item{dABO}{A character value with donor's ABO blood group (\verb{r env$valid.blood.groups}).}

\item{dA}{A two elements character vector donor's HLA-A typing.}

\item{dB}{A two elements character vector donor's HLA-B typing.}

\item{dDR}{A two elements character vector donor's HLA-DR typing.}

\item{donor.age}{A numeric value with donor's age.}

\item{df.abs}{A data frame with candidates' antibodies.}

\item{data}{A data frame containing demographics and medical information
for a group of waitlisted transplant candidates.}

\item{n}{A positive integer to slice the first candidates.}

\item{q2}{A numerical value for the median of candidates' waiting list (\verb{r env$q.minimum} - \verb{r env$q.maximum}).}

\item{q3}{A numerical value for the third quartile of candidates' waiting list (\verb{r env$q.minimum} - \verb{r env$q.maximum}).}

\item{uj.matx}{A 6 by 6 matrix obtained from \code{uj_matx()}}

\item{check.validity}{Logical to decide whether to validate input arguments.}
}
\value{
An ordered data frame with a column \code{pts} (points from utility-justice matrix), \code{SP}, \code{HI} and \code{mmHLA}.
}
\description{
An algorithm to apply on deceased donor's Kidney allocation for transplantation.
Ordering of wait listed candidates for a given donor and according to and utility-justice matrix.
}
\examples{
\dontrun{
eqm_v0 <- function(iso = TRUE, dABO = "O" , 
dA = c("1","2"), dB = c("15","44"), dDR = c("1","4") , 
donor.age = 60 , df.abs = cabs , data = candidates , 
n = 2 , q2 = 60 , q3 = 80 , uj.matx = uj_matx(), 
check.validity = FALSE)
}
}
